+++ run.py
@@ -1,0 +1,5 @@
+from app import app
+
+# An Alternative Way to create a debug mode
+if __name__ == '__main__':
+    app.run(debug=True)
